refactor: delete @handlebars/parser and old multipass pipeline
Remove the entire old parsing pipeline that is now replaced by the
Rust/WASM pest.rs parser:

Deleted packages:
- packages/@handlebars/parser/ (Jison-generated Handlebars parser)

Deleted files from @glimmer/syntax:
- lib/parser.ts (abstract Parser base class using simple-html-tokenizer)
- lib/parser/handlebars-node-visitors.ts (HBS AST → ASTv1 visitor)
- lib/parser/preprocess-rust.ts (merged into tokenizer-event-handlers.ts)
- lib/v1/handlebars-ast.ts (intermediate HBS AST types)
- lib/hbs-parser/parser.js (generated Jison parser)

Removed dependencies:
- @handlebars/parser (workspace dependency)
- simple-html-tokenizer (from @glimmer/syntax — still used by test helpers)

Updated configs:
- rollup.config.mjs: removed hidden dependency entries
- eslint.config.mjs: removed @handlebars eslint rules
- CI workflows: removed @handlebars/parser type checking and tests
- build-constraints.md: updated external dependency docs

The tokenizer-event-handlers.ts file now contains the Rust parser
wrapper directly, keeping the same export path so all existing
consumers (index.ts, get-template-locals.ts, v2/normalize.ts) work
without import changes.
