+++ package.json
@@ -55,7 +55,6 @@
     "test:browserstack": "node bin/run-browserstack-tests.js",
     "test:wip": "vite build --mode development --minify false && testem ci",
     "type-check:internals": "tsc --noEmit",
-    "type-check:handlebars": "tsc --noEmit --project packages/@handlebars/parser/tsconfig.json",
     "type-check:types": "tsc --noEmit --project type-tests",
     "type-check": "npm-run-all type-check:*"
   },
@@ -76,8 +75,7 @@
     "inflection": "^2.0.1",
     "route-recognizer": "^0.3.4",
     "semver": "^7.5.2",
-    "silent-error": "^1.1.1",
-    "simple-html-tokenizer": "^0.5.11"
+    "silent-error": "^1.1.1"
   },
   "devDependencies": {
     "@aws-sdk/client-s3": "^3.731.0",
