+++ packages/@glimmer/syntax/package.json
@@ -35,9 +35,7 @@
   "dependencies": {
     "@glimmer/interfaces": "workspace:*",
     "@glimmer/util": "workspace:*",
-    "@glimmer/wire-format": "workspace:*",
-    "@handlebars/parser": "workspace:*",
-    "simple-html-tokenizer": "^0.5.11"
+    "@glimmer/wire-format": "workspace:*"
   },
   "devDependencies": {
     "@glimmer/debug-util": "workspace:*",
