--- packages/@handlebars/parser/CHANGELOG.md
@@ -1,80 +1,0 @@
-# Changelog
-
-## Release (2025-11-29)
-
-* @handlebars/parser 2.2.2 (patch)
-
-#### :bug: Bug Fix
-* `@handlebars/parser`
-  * [#27](https://github.com/handlebars-lang/handlebars-parser/pull/27) Fix commonjs build ([@kaermorchen](https://github.com/kaermorchen))
-
-#### Committers: 1
-- Stanislav Romanov ([@kaermorchen](https://github.com/kaermorchen))
-
-## Release (2025-08-01)
-
-* @handlebars/parser 2.2.1 (patch)
-
-#### :bug: Bug Fix
-* `@handlebars/parser`
-  * [#24](https://github.com/handlebars-lang/handlebars-parser/pull/24) fix node engine and create a test matrix in CI ([@mansona](https://github.com/mansona))
-
-#### Committers: 1
-- Chris Manson ([@mansona](https://github.com/mansona))
-
-## Release (2025-03-19)
-
-* @handlebars/parser 2.2.0 (minor)
-
-#### :rocket: Enhancement
-* `@handlebars/parser`
-  * [#15](https://github.com/handlebars-lang/handlebars-parser/pull/15) Implement hash and array literal syntax ([@wycats](https://github.com/wycats))
-  * [#14](https://github.com/handlebars-lang/handlebars-parser/pull/14) Add support for a `#` prefix in path segments ([@wycats](https://github.com/wycats))
-
-#### :bug: Bug Fix
-* `@handlebars/parser`
-  * [#6](https://github.com/handlebars-lang/handlebars-parser/pull/6) [bugfix] Don't emit `parts: [undefined]` for `{{this}}` ([@dfreeman](https://github.com/dfreeman))
-
-#### :house: Internal
-* `@handlebars/parser`
-  * [#17](https://github.com/handlebars-lang/handlebars-parser/pull/17) Add release plan, remove release-it ([@NullVoxPopuli](https://github.com/NullVoxPopuli))
-
-#### Committers: 3
-- Dan Freeman ([@dfreeman](https://github.com/dfreeman))
-- Yehuda Katz ([@wycats](https://github.com/wycats))
-- [@NullVoxPopuli](https://github.com/NullVoxPopuli)
-
-## v2.1.0 (2021-02-22)
-
-#### :rocket: Enhancement
-* [#4](https://github.com/handlebars-lang/handlebars-parser/pull/4) [FEATURE] Allows SubExpressions to be PathExpression roots ([@pzuraq](https://github.com/pzuraq))
-
-#### Committers: 1
-- Chris Garrett ([@pzuraq](https://github.com/pzuraq))
-
-## v2.0.0 (2020-12-09)
-
-#### :boom: Breaking Change
-* [#3](https://github.com/handlebars-lang/handlebars-parser/pull/3) Make sub-expressions callable ([@pzuraq](https://github.com/pzuraq))
-
-#### Committers: 1
-- Chris Garrett ([@pzuraq](https://github.com/pzuraq))
-
-## v1.1.0 (2020-09-18)
-
-#### :rocket: Enhancement
-* [#2](https://github.com/handlebars-lang/handlebars-parser/pull/2) [FEAT] Adds types ([@pzuraq](https://github.com/pzuraq))
-
-#### Committers: 1
-- Chris Garrett ([@pzuraq](https://github.com/pzuraq))
-
-
-## v1.0.0 (2020-09-14)
-
-#### :rocket: Enhancement
-* [#1](https://github.com/handlebars-lang/handlebars-parser/pull/1) [FEAT] Adds initial setup, lint, ci ([@pzuraq](https://github.com/pzuraq))
-
-#### Committers: 1
-- Chris Garrett ([@pzuraq](https://github.com/pzuraq))
-
-
