--- packages/@handlebars/parser/README.md
@@ -1,4 +1,0 @@
-# Handlebars Parser
-
-The official Handlebars.js parser. This package contains the definition for
-for the Handlebars language AST, and the parser for parsing into that AST.
