--- packages/@handlebars/parser/lib/index.js
@@ -1,6 +1,0 @@
-export { default as Visitor } from './visitor.js';
-export { default as WhitespaceControl } from './whitespace-control.js';
-export { default as parser } from './parser.js';
-export { default as Exception } from './exception.js';
-export { print, PrintVisitor } from './printer.js';
-export { parse, parseWithoutProcessing } from './parse.js';
