--- packages/@handlebars/parser/src/parser-suffix.js
@@ -1,1 +1,0 @@
-export default parser;
