--- packages/@handlebars/parser/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    // Compilation Configuration
-    "target": "es2017",
-    "inlineSources": true,
-    "inlineSourceMap": true,
-    "outDir": "dist",
-    "rootDir": "lib",
-    "esModuleInterop": true,
-    "moduleResolution": "bundler",
-    "verbatimModuleSyntax": true,
-    // Enhance Strictness
-    "strict": true,
-    "skipLibCheck": true,
-    "suppressImplicitAnyIndexErrors": false,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "newLine": "LF",
-    "allowJs": true
-  },
-  "include": ["lib/**/*.js", "lib/**/*.d.ts"],
-  "exclude": ["lib/parser.js", "dist", "node_modules", ".vscode"]
-}
