+++ rollup.config.mjs
@@ -216,9 +216,6 @@ function packages() {
       // "exposedDependencies" since they used to actually be dependencies.
       '@glimmer-workspace/**',
       '@glimmer/**',
-
-      // @handlebars/parser is a hidden dependency, not an explicit entrypoint
-      '@handlebars/**',
     ],
     cwd: 'packages',
   });
@@ -279,11 +276,6 @@ export function exposedDependencies() {
 // expose to consumers
 export function hiddenDependencies() {
   return {
-    'simple-html-tokenizer': entrypoint(
-      findFromProject('@glimmer/syntax', 'simple-html-tokenizer'),
-      'module'
-    ).path,
-    '@handlebars/parser': resolve(packageCache.appRoot, 'packages/@handlebars/parser/lib/index.js'),
     ...walkGlimmerDeps(['@glimmer/compiler']),
     'decorator-transforms/runtime': resolve(
       findFromProject('decorator-transforms').root,
